lr_scheduler bug fix and cleanup comments

+++ reward_model_trainer.py
@@ -40,17 +40,6 @@ class PickScoreTrainer:
         
         self.accelerator = instantiate_with_cfg(cfg.accelerator)
         print("\ninstantiate_with_cfg accelerator type", self.accelerator, type(self.accelerator))
-        # breakpoint()
-        # self.accelerator = Accelerator(
-        #     gradient_accumulation_steps=cfg.accelerator.gradient_accumulation_steps,
-        #     mixed_precision=cfg.accelerator.mixed_precision,
-        #     log_with=cfg.accelerator.log_with,
-        #     project_dir=cfg.accelerator.output_dir,
-        #     dynamo_backend=cfg.accelerator.dynamo_backend,
-        # )
-        # print("accelerator type", self.accelerator, type(self.accelerator))
-        # breakpoint()
-        # print("\nDistributed type: ", self.accelerator.distributed_type)
 
         if cfg.debug.activate and self.accelerator.is_main_process:
             import pydevd_pycharm
@@ -82,14 +71,7 @@ class PickScoreTrainer:
         print("model type (after prepare): ", type(self.model))
 
         self.accelerator.load_state_if_needed()
-
-        # self.accelerator.recalc_train_length_after_prepare(len(split2dataloader[cfg.dataset.train_split_name]))
-
         self.accelerator.init_training(cfg)
-        # if self.accelerator.is_main_process:
-        #     yaml = OmegaConf.to_yaml(cfg.accelerator, resolve=True, sort_keys=True)
-        #     log_cfg = _flatten_dict(OmegaConf.create(yaml))
-        #     self.accelerator.accelerator.init_trackers(cfg.accelerator.project_name, log_cfg)
 
         processor_name_or_path = "laion/CLIP-ViT-H-14-laion2B-s32B-b79K"
         self.processor = AutoProcessor.from_pretrained(processor_name_or_path)
@@ -102,22 +84,24 @@ class PickScoreTrainer:
         self.query_generator = QueryGenerator()
 
         # Create files to save images and dataset to
-        self.image_save_path = os.path.join(cfg.output_dir, "image_data")
-        date_and_time = datetime.datetime.now().strftime("%Y%m%d-%H%M%S")
-        self.dataset_save_path = os.path.join(cfg.output_dir, f"preference_data/{date_and_time}")
-        if os.path.exists(self.image_save_path):
-            shutil.rmtree(self.image_save_path)
-        os.makedirs(self.image_save_path, exist_ok=False)
-        os.makedirs(self.dataset_save_path, exist_ok=False)
+        if self.accelerator.is_main_process:
+            self.image_save_path = os.path.join(cfg.output_dir, "image_data")
+            date_and_time = datetime.datetime.now().strftime("%Y%m%d-%H%M%S")
+            self.dataset_save_path = os.path.join(cfg.output_dir, f"preference_data/{date_and_time}")
+            if os.path.exists(self.image_save_path):
+                shutil.rmtree(self.image_save_path)
+            os.makedirs(self.image_save_path, exist_ok=False)
+            os.makedirs(self.dataset_save_path, exist_ok=False)
 
+        self.split2dataloader = None
         self.cfg = cfg
 
-    def evaluate(self, split2dataloader):
+    def evaluate(self, ):
         self.model.eval()
         end_of_train_dataloader = self.accelerator.gradient_state.end_of_dataloader
         print(f"*** Evaluating {self.cfg.dataset.valid_split_name} ***")
         # logger.info(f"*** Evaluating {cfg.dataset.valid_split_name} ***")
-        metrics = self.task.evaluate(self.model, self.criterion, split2dataloader[self.cfg.dataset.valid_split_name])
+        metrics = self.task.evaluate(self.model, self.criterion, self.split2dataloader[self.cfg.dataset.valid_split_name])
         self.accelerator.update_metrics(metrics)
         # accelerator.gradient_state.end_of_dataloader = end_of_train_dataloader
 
@@ -183,7 +167,7 @@ class PickScoreTrainer:
         OmegaConf.save(cfg, yaml_path, resolve=True) # TODO
 
         # if not os.path.exists(yaml_path):
-        #     OmegaConf.save(cfg, yaml_path, resolve=True) # TODO
+        #     OmegaConf.save(cfg, yaml_path, resolve=True)
         # with open(yaml_path) as f:
         #     existing_config = f.read()
         # if existing_config != OmegaConf.to_yaml(cfg, resolve=True):
@@ -195,23 +179,27 @@ class PickScoreTrainer:
         #######################################################
         ########### Active Query and Dataset Update ###########
         #######################################################
-        queries = self.query_generator.generate_queries(
-            images=image_batch,
-            query_algorithm="random", # TODO - add to config
-            n_queries=3, # TODO - add to config
-        )
+        if self.accelerator.is_main_process:
+            queries = self.query_generator.generate_queries(
+                images=image_batch,
+                query_algorithm="random", # TODO - add to config
+                n_queries=3, # TODO - add to config
+            )
 
-        # Collect preferences
-        self.feedback_interface.reset_dataset() # clear data from previous iteration
-        self.feedback_interface.query_batch(
-            prompt=prompt,
-            image_batch=image_batch,
-            query_indices=queries,
-        )
+            # Collect preferences
+            self.feedback_interface.reset_dataset() # clear data from previous iteration
+            self.feedback_interface.query_batch(
+                prompt=prompt,
+                image_batch=image_batch,
+                query_indices=queries,
+            )
 
-        # Save new dataset and reinitialize dataloaders - TODO
-        self.feedback_interface.save_dataset(dataset_save_path=os.path.join(self.dataset_save_path, f"epoch{epoch}.parquet"))
-        self.feedback_interface.save_dataset(dataset_save_path="rl4dgm/my_dataset/my_dataset_train.parquet") 
+            # Save new dataset and reinitialize dataloaders - TODO
+            self.feedback_interface.save_dataset(dataset_save_path=os.path.join(self.dataset_save_path, f"epoch{epoch}.parquet"))
+            self.feedback_interface.save_dataset(dataset_save_path="rl4dgm/my_dataset/my_dataset_train.parquet") 
+
+        # wait for the main process to update the training dataset
+        self.accelerator.wait_for_everyone()
 
         # Initialize trainloader from new dataset
         trainloader = self.load_dataloaders(self.cfg.dataset, split="train")
@@ -220,21 +208,13 @@ class PickScoreTrainer:
         validloader = self.load_dataloaders(self.cfg.dataset, split="validation")
 
         trainloader, validloader = self.accelerator.prepare(trainloader, validloader)
-        split2dataloader = {
+        self.split2dataloader = {
             "train" : trainloader,
             "validation" : validloader,
         }
-        dataloaders = list(split2dataloader.values())
+        dataloaders = list(self.split2dataloader.values())
 
-        self.accelerator.recalc_train_length_after_prepare(len(split2dataloader[self.cfg.dataset.train_split_name]))
-        # num_batches = len(split2dataloader[self.cfg.dataset.train_split_name])
-        # num_update_steps_per_epoch = math.ceil(num_batches / self.cfg.accelerator.gradient_accumulation_steps)
-        # if self.cfg.accelerator.max_steps is None:
-        #     self.cfg.accelerator.max_steps = self.cfg.accelerator.num_epochs * num_update_steps_per_epoch
-        # self.num_update_steps_per_epoch = num_update_steps_per_epoch
-        # self.num_steps_per_epoch = num_batches
-        # self.cfg.accelerator.num_epochs = math.ceil(self.cfg.accelerator.max_steps / num_update_steps_per_epoch)
-
+        self.accelerator.recalc_train_length_after_prepare(len(self.split2dataloader[self.cfg.dataset.train_split_name]))
 
         #######################################################
         ################ Reward Model Training ################
@@ -243,20 +223,16 @@ class PickScoreTrainer:
 
         train_loss, lr = 0.0, 0.0
         for inner_epoch in range(self.cfg.accelerator.num_epochs): # TODO config should have something like reward epochs per loop
-            print("Epoch ", inner_epoch)
-            for step, batch in enumerate(split2dataloader[self.cfg.dataset.train_split_name]):
-                # TODO
+            # print("Epoch ", inner_epoch)
+            for step, batch in enumerate(self.split2dataloader[self.cfg.dataset.train_split_name]):
                 if self.accelerator.should_skip(inner_epoch, step):
                     self.accelerator.update_progbar_step()
                     continue
 
                 if self.accelerator.should_eval():
-                    # pass
-                    # TODO where should eval() be defined? 
-                    # validation dataset can be accumulation of all previous human feedbacks?
-                    self.evaluate(split2dataloader)
+                    # TODO - validation dataset can be accumulation of all previous human feedbacks?
+                    self.evaluate()
 
-                # TODO
                 if self.accelerator.should_save():
                     self.accelerator.save_checkpoint()
 
@@ -270,10 +246,6 @@ class PickScoreTrainer:
                     if self.accelerator.sync_gradients:
                         self.accelerator.clip_grad_norm_(self.model.parameters())
 
-                    # NOTE: default optimizer and lr_scheduler are dummy, which gets wrapped with DeepSpeedOptimizerWrapper and DeepSpeedSchedulerWrapper.
-                    # wrapper's step functions do nothing (pass)
-                    # TODO - is this necessary now that deepspeed is removed?
-
                     self.optimizer.step()
                     self.lr_scheduler.step()
                     self.optimizer.zero_grad()
@@ -281,14 +253,13 @@ class PickScoreTrainer:
                 # train_loss += avg_loss / self.accelerator.cfg.gradient_accumulation_steps
                 train_loss += avg_loss / self.cfg.accelerator.gradient_accumulation_steps
 
-                # TODO
                 if self.accelerator.sync_gradients:
                     self.accelerator.update_global_step(train_loss)
                     train_loss = 0.0
 
                 if self.accelerator.global_step > 0:
                     try:
-                        lr = lr_scheduler.get_last_lr()[0]
+                        lr = self.lr_scheduler.get_last_lr()[0]
                     except:
                         print("get_last_lr exception. Setting lr=0.0")
                         lr = 0.0
@@ -297,7 +268,7 @@ class PickScoreTrainer:
                 self.accelerator.lr = lr
 
                 if self.accelerator.should_end():
-                    evaluate()
+                    self.evaluate()
                     self.accelerator.save_checkpoint()
                     break
